Separando o footer e o header para um componente
