+++ frontend/src/app/app.component.css
@@ -1,46 +1,3 @@
-.logo{
-  display: flex;
-  align-items: center;
-  padding: 5px;
-}
-
-.dash{
-  background-color: var(--color-blue);
-  width: 3px;
-  height: 50px;
-  margin: 0 5px;
-}
-
-.title-logo{
-  display: flex;
-  flex-direction: column;
-  font-size: 24px;
-  line-height: 20px;
-  color: var(--color-white);
-}
-
-.navbar{
-  background-color: var(--color-black);
-  box-shadow: 0 2px 7px 0 rgba(0,0,0,.5);
-}
-
-.perfil button{
-  background-color: var(--color-blue);
-  color: var(--color-white);
-}
-
-.perfil button:hover{
-  background-color: var(--color-white);
-  color: var(--color-blue);
-}
-
-footer {
-  margin-top: 30px;
-  color: var(--color-white);
-  background-color: var(--color-black);
-  box-shadow: 0 -2px 7px 0 rgba(0,0,0,.5);
-}
-
 .content{
   min-height: calc(100vh - 76px - 56px - 30px);
   height: auto;
