+++ frontend/src/app/components/template/footer/footer.component.css
@@ -1,0 +1,6 @@
+footer {
+  margin-top: 30px;
+  color: var(--color-white);
+  background-color: var(--color-black);
+  box-shadow: 0 -2px 7px 0 rgba(0,0,0,.5);
+}
