+++ frontend/src/app/components/template/footer/footer.component.html
@@ -1,0 +1,5 @@
+<footer class="text-center text-font">
+  <div class="text-center p-3" style="background-color: rgba(0, 0, 0, 0.2);">
+    © 2022 Copyright: Desenvolvido pelo Gabriel Borges
+  </div>
+</footer>
