+++ frontend/src/app/components/template/header/header.component.html
@@ -1,0 +1,18 @@
+<header>
+  <nav class="navbar">
+    <div class="container">
+      <div class="logo">
+        <img src="../assets/imgs/scheduling_icon_white.png" alt="Imagem de um calendário com relógio a frente" width="50" height="50">
+        <div class="dash"></div>
+        <div class="title-logo title-font">
+          <span>Event</span> <span>Scheduling</span>
+        </div>
+      </div>
+        <div [ngClass]="'perfil'" [ngStyle]="{display: getVisibility()}">
+            <button type="button" class="btn" (click)="openModal()">
+              <i class="bi bi-box-arrow-right"></i>
+            </button>
+        </div>
+    </div>
+  </nav>
+</header>
